🐛 fix(eval): enforce judge scoring semantics
